Get this library compiling on Windows
