+++ clamd.go
@@ -20,7 +20,6 @@ import (
 	"os"
 	"regexp"
 	"strings"
-	"syscall"
 	"time"
 
 	"github.com/baruwa-enterprise/clamd/protocol"
@@ -450,34 +449,6 @@ func (c *Client) instreamScan(tc *textproto.Conn, conn net.Conn, p string) (err 
 	defer f.Close()
 
 	if err = c.streamCmd(tc, protocol.Instream, f, conn); err != nil {
-		return
-	}
-
-	return
-}
-
-func (c *Client) fildesScan(tc *textproto.Conn, conn net.Conn, p string) (err error) {
-	var f *os.File
-	var vf *os.File
-
-	fmt.Fprintf(tc.W, "n%s\n", protocol.Fildes)
-	tc.W.Flush()
-
-	if f, err = os.Open(p); err != nil {
-		return
-	}
-	defer f.Close()
-
-	s := conn.(*net.UnixConn)
-	if vf, err = s.File(); err != nil {
-		return
-	}
-	sock := int(vf.Fd())
-	defer vf.Close()
-
-	fds := []int{int(f.Fd())}
-	rights := syscall.UnixRights(fds...)
-	if err = syscall.Sendmsg(sock, nil, rights, nil, 0); err != nil {
 		return
 	}
